# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API MoliCasillas",
        "description": "Documentación de la API para el Sistema de Casilla Electrónica MoliCasillas",
        "contact": {
            "email": "admin@molicasillas.gob.pe"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://molicasillas.test",
            "description": "Servidor de API MoliCasillas"
        }
    ],
    "paths": {
        "/api/v1/casilla/messages/{mensajeId}/attachments": {
            "get": {
                "tags": [
                    "Adjuntos"
                ],
                "summary": "Listar adjuntos de un mensaje",
                "operationId": "9a650781243f3b4deea9126432019c18",
                "parameters": [
                    {
                        "name": "mensajeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de adjuntos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Adjunto"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Mensaje no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/casilla/attachments/{id}/download": {
            "get": {
                "tags": [
                    "Adjuntos"
                ],
                "summary": "Descargar adjunto",
                "operationId": "a87048fc6a7dbca25dc9b9b9071a5e8b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Stream de descarga del archivo",
                        "content": {
                            "application/octet-stream": {}
                        }
                    },
                    "404": {
                        "description": "Adjunto no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/auth/login": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Iniciar sesión con ID externo",
                "operationId": "a7997e7c1e3bfed64d4e56de82415931",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "usuario_auth_id",
                                    "tipo_documento",
                                    "numero_documento",
                                    "nombre_mostrar"
                                ],
                                "properties": {
                                    "usuario_auth_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "11111111-1111-1111-1111-111111111111"
                                    },
                                    "tipo_documento": {
                                        "type": "string",
                                        "enum": [
                                            "DNI",
                                            "CE",
                                            "RUC"
                                        ],
                                        "example": "DNI"
                                    },
                                    "numero_documento": {
                                        "type": "string",
                                        "example": "12345678"
                                    },
                                    "nombre_mostrar": {
                                        "type": "string",
                                        "example": "Juan Pérez"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "juan@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inicio de sesión exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "token": {
                                                    "type": "string",
                                                    "example": "1|AbCdEf..."
                                                },
                                                "tipo_token": {
                                                    "type": "string",
                                                    "example": "Bearer"
                                                },
                                                "cuenta": {
                                                    "$ref": "#/components/schemas/Casilla"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/v1/auth/logout": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Cerrar sesión",
                "operationId": "f6b8957b4c0b25636613a9f0512a8ba9",
                "responses": {
                    "200": {
                        "description": "Sesión cerrada correctamente"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/me": {
            "get": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Obtener usuario autenticado",
                "operationId": "13b43df400e475a157538f73617b8c3d",
                "responses": {
                    "200": {
                        "description": "Datos del usuario actual",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Casilla"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Adjunto": {
                "title": "Adjunto",
                "description": "Adjunto de mensaje",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "mensaje_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "nombre_archivo": {
                        "type": "string"
                    },
                    "tipo_mime": {
                        "type": "string"
                    },
                    "tamano_bytes": {
                        "type": "integer"
                    },
                    "subido_en": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Casilla": {
                "title": "Casilla",
                "description": "Cuenta de usuario de casilla",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "usuario_auth_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "tipo_documento": {
                        "type": "string",
                        "enum": [
                            "DNI",
                            "CE",
                            "RUC"
                        ]
                    },
                    "numero_documento": {
                        "type": "string"
                    },
                    "nombre_mostrar": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "estado": {
                        "type": "string",
                        "enum": [
                            "ACTIVO",
                            "SUSPENDIDO"
                        ]
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}